+++ includes/algorithms.h
@@ -1,3 +1,3 @@
-void shell_sort(int* arr, int size, int reverse_bit); // shell sorting algorithm function
-void insertion_sort(int* arr, int size, int reverse_bit); // insertion sorting algorithm function
-int bsearch_more(int* arr, int size, int key); // binary searching function for searching min index of element, that more than key
+void shell_sort(long long* arr, long long size, short reverse_bit); // shell sorting algorithm function
+void insertion_sort(long long* arr, long long size, short reverse_bit); // insertion sorting algorithm function
+long long bsearch_more(long long* arr, long long size, long long key); // binary searching function for searching min index of element, that more than key
